Send chime to service
