+++ app/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     id("com.android.application")
     id("org.jetbrains.kotlin.android")
     id("app.cash.sqldelight")
+    id("kotlin-parcelize")
 }
 
 android {
