+++ app/src/main/java/org/simonolander/horloge/model/Chime.kt
@@ -1,12 +1,15 @@
 package org.simonolander.horloge.model
 
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
 import java.util.UUID
 
+@Parcelize
 data class Chime(
     val id: String,
     val name: String,
     val beats: List<Beat>,
-) {
+) : Parcelable {
     companion object {
         fun randomId() = UUID.randomUUID().toString()
     }
